requirements.txt
amqp==2.1.4
appnope==0.1.0
billiard==3.5.0.2
boltons==16.1.1
brotlipy==0.6.0
celery==4.0.2
certifi==2016.9.26
cffi==1.9.1
coverage==4.3.1
decorator==4.0.10
dj-database-url==0.4.1
Django==1.10.4
django-log-request-id==1.3.2
django-model-utils==2.6
django-naomi==0.8
django-webpack-loader==0.4.1
django.js-vinta==0.8.2.dev2
djangorestframework==3.5.3
fixmydjango==0.1
flake8==3.2.1
gunicorn==19.6.0
ipython==5.1.0
ipython-genutils==0.1.0
kombu==4.0.2
mccabe==0.5.3
model-mommy==1.3.1
opbeat==3.5.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
pycodestyle==2.2.0
pycparser==2.17
pyflakes==1.3.0
Pygments==2.1.3
python-decouple==3.0
pytz==2016.10
redis==2.10.5
requests==2.9.1
simplegeneric==0.8.1
six==1.10.0
termcolor==1.1.0
traitlets==4.3.1
urllib3==1.19.1
vine==1.1.3
wcwidth==0.1.7
wheel==0.24.0
whitenoise==3.2.3
